api/requirements.txt
fastapi
uvicorn
python-nmap
redis
rq
pydantic